+++ WWW/Library/Implementation/HTRules.c
@@ -473,9 +473,9 @@ PUBLIC int  HTSetConfiguration ARGS1(
 			    &quality, &secs, &secs_per_byte, &maxbytes);
 	else status = 0;
 	HTSetPresentation(word2, word3,
-		    status >= 1 ? quality		: (float) 1.0,
-		    status >= 2 ? secs			: (float) 0.0,
-		    status >= 3 ? secs_per_byte 	: (float) 0.0,
+		    status >= 1 ? quality		: 1.0,
+		    status >= 2 ? secs			: 0.0,
+		    status >= 3 ? secs_per_byte 	: 0.0,
 		    status >= 4 ? maxbytes		: 0 );
 
     } else if (0==strncasecomp(word1, "htbin", 5) ||
